Introduce `EventSubscriber` interface
