+++ server/src/test/java/io/spine/server/event/given/EventBusTestEnv.java
@@ -47,7 +47,7 @@ import io.spine.server.event.EventBusTest;
 import io.spine.server.event.EventDispatcher;
 import io.spine.server.event.EventEnricher;
 import io.spine.server.event.EventStreamQuery;
-import io.spine.server.event.EventSubscriber;
+import io.spine.server.event.AbstractEventSubscriber;
 import io.spine.server.tenant.TenantAwareOperation;
 import io.spine.test.event.EBProjectArchived;
 import io.spine.test.event.EBProjectCreated;
@@ -277,7 +277,7 @@ public class EventBusTestEnv {
      *
      * <p>Can be used for the event to get pass the {@link io.spine.server.bus.DeadMessageFilter}.
      */
-    public static class EBProjectCreatedNoOpSubscriber extends EventSubscriber {
+    public static class EBProjectCreatedNoOpSubscriber extends AbstractEventSubscriber {
 
         @Subscribe
         public void on(EBProjectCreated message, EventContext context) {
@@ -285,7 +285,7 @@ public class EventBusTestEnv {
         }
     }
 
-    public static class EBProjectArchivedSubscriber extends EventSubscriber {
+    public static class EBProjectArchivedSubscriber extends AbstractEventSubscriber {
 
         private Message eventMessage;
 
@@ -299,7 +299,7 @@ public class EventBusTestEnv {
         }
     }
 
-    public static class ProjectCreatedSubscriber extends EventSubscriber {
+    public static class ProjectCreatedSubscriber extends AbstractEventSubscriber {
 
         private Message eventMessage;
         private EventContext eventContext;
@@ -323,7 +323,7 @@ public class EventBusTestEnv {
      * {@link EBTaskAdded} subscriber that does nothing. Can be used for the event to get pass the
      * {@link io.spine.server.bus.DeadMessageFilter}.
      */
-    public static class EBTaskAddedNoOpSubscriber extends EventSubscriber {
+    public static class EBTaskAddedNoOpSubscriber extends AbstractEventSubscriber {
 
         @Subscribe
         public void on(EBTaskAdded message, EventContext context) {
